Services

+++ AutoFixPlatform/src/main/java/com/tingesoEv1/AutoFixPlatform/repositories/BonusRepository.java
@@ -9,5 +9,5 @@ public interface BonusRepository extends JpaRepository<BonusEntity, Long> {
     List<BonusEntity> findByBrand(String brand);
     List<BonusEntity> findByYear(int year);
     List<BonusEntity> findByYearAndMonth(Integer year, Integer month);
-    List<BonusEntity> findByAmountLessThan(int amount);
+    List<BonusEntity> findByAmountLessThan(double amount);
 }
