+++ AutoFixPlatform/src/main/java/com/tingesoEv1/AutoFixPlatform/services/BonusService.java
@@ -37,7 +37,7 @@ public class BonusService {
         return bonusRepository.findByYearAndMonth(year, month);
     }
 
-    public List<BonusEntity> getBonusByAmount(int amount) {
+    public List<BonusEntity> getBonusByAmount(double amount) {
         return bonusRepository.findByAmountLessThan(amount);
     }
